restructured file structure to better understand the flow of data and improve readability.
